Prevent recursive import crash on Linux/OSX (#117)
Our naming convention happens to overlap with a wildcard include
statement in the vanilla NotificationPanel.lua

This issue only affects Linux/OSX (I think...).
Which is bizarre, since typically you'd expect case sensitivity
bugs to be caused by /stricter/ file name matching,
not less strict!

I suspect Aspyr wrapped the include statement so that it converts the
filename to lowercase before passing it to the function. If that's the
case, the wildcard import, specifically may be case-sensitive on the
Windows platform.

Again... bizarre! But that's the only explanation I can cook up for this
unexpected behavior.

Here's my trace if you're curious:
stack traceback:
	[C]: in function 'lInclude'
	C:\Emu\AppAssets\Base\Assets\ui\panels\notificationpanel.lua:3: in function '(main chunk)'
	[C]: in function 'lInclude'
	\steam\SteamLibrary\steamapps\workshop\content\289070\2115302648\Assets\UI\Panels\notificationpanel_CQUI_basegame.lua:4: in function '(main chunk)'
	[C]: in function 'lInclude'
	C:\Emu\AppAssets\Base\Assets\ui\panels\notificationpanel.lua:1895: in function '(main chunk)'
	[C]: in function 'lInclude'
	\steam\SteamLibrary\steamapps\workshop\content\289070\2115302648\Assets\UI\Panels\notificationpanel_CQUI_basegame.lua:4: in function '(main chunk)'
	[C]: in function 'lInclude'
	C:\Emu\AppAssets\Base\Assets\ui\panels\notificationpanel.lua:1895: in function '(main chunk)'
Lua callstack:

The recursion is coming from this line in NotificationPanel.lua:
include("NotificationPanel_", true);