ActionPanel: Fix Policy Reminder
Modifies CQUI_CheckPolicyCanBeChanged function to be more dynamic. This avoids triggering on the first turn and on future civic like before, but also avoids triggering on civics that have no policy cards, such as the Tier 4 Government civics.

Also moves the statement that resets the status of the policy reminder so it no longer depends on Auto Unit Cycle being off.